Move atari/mujoco helpers into package code
Move mujoco_env.py from examples/ into tianshou/env/mujoco/ and add
__init__.py for both tianshou/env/atari/ and tianshou/env/mujoco/ to
enable clean imports like `from tianshou.env.mujoco import make_mujoco_env`.

The original examples/mujoco/mujoco_env.py is preserved as a backward-
compatible re-export shim.

Closes #988
