+++ examples/atari/atari_dqn_hl.py
@@ -5,10 +5,11 @@ from typing import Literal
 
 from sensai.util import logging
 
-from tianshou.env.atari.atari_network import (
+from tianshou.env.atari import (
+    AtariEnvFactory,
+    AtariEpochStopCallback,
     IntermediateModuleFactoryAtariDQN,
 )
-from tianshou.env.atari.atari_wrapper import AtariEnvFactory, AtariEpochStopCallback
 from tianshou.highlevel.config import OffPolicyTrainingConfig
 from tianshou.highlevel.experiment import (
     DQNExperimentBuilder,
