+++ examples/atari/atari_fqf.py
@@ -14,8 +14,7 @@ from tianshou.algorithm.algorithm_base import Algorithm
 from tianshou.algorithm.modelfree.fqf import FQFPolicy
 from tianshou.algorithm.optim import AdamOptimizerFactory, RMSpropOptimizerFactory
 from tianshou.data import Collector, CollectStats, VectorReplayBuffer
-from tianshou.env.atari.atari_network import DQNet
-from tianshou.env.atari.atari_wrapper import make_atari_env
+from tianshou.env.atari import DQNet, make_atari_env
 from tianshou.highlevel.logger import LoggerFactoryDefault
 from tianshou.trainer import OffPolicyTrainerParams
 from tianshou.utils.net.discrete import FractionProposalNetwork, FullQuantileFunction
