+++ examples/mujoco/mujoco_ddpg_hl.py
@@ -5,7 +5,7 @@ from typing import Literal
 
 from sensai.util import logging
 
-from examples.mujoco.mujoco_env import MujocoEnvFactory
+from tianshou.env.mujoco import MujocoEnvFactory
 from tianshou.highlevel.config import OffPolicyTrainingConfig
 from tianshou.highlevel.experiment import (
     DDPGExperimentBuilder,
