+++ examples/mujoco/mujoco_td3.py
@@ -6,7 +6,6 @@ import pprint
 
 import numpy as np
 import torch
-from mujoco_env import make_mujoco_env
 from sensai.util import logging
 
 from tianshou.algorithm import TD3
@@ -14,6 +13,7 @@ from tianshou.algorithm.algorithm_base import Algorithm
 from tianshou.algorithm.modelfree.ddpg import ContinuousDeterministicPolicy
 from tianshou.algorithm.optim import AdamOptimizerFactory
 from tianshou.data import Collector, CollectStats, ReplayBuffer, VectorReplayBuffer
+from tianshou.env.mujoco import make_mujoco_env
 from tianshou.exploration import GaussianNoise
 from tianshou.highlevel.logger import LoggerFactoryDefault
 from tianshou.trainer import OffPolicyTrainerParams
