+++ examples/offline/atari_bcq.py
@@ -17,8 +17,7 @@ from tianshou.algorithm.algorithm_base import Algorithm
 from tianshou.algorithm.imitation.discrete_bcq import DiscreteBCQPolicy
 from tianshou.algorithm.optim import AdamOptimizerFactory
 from tianshou.data import Collector, CollectStats, VectorReplayBuffer
-from tianshou.env.atari.atari_network import DQNet
-from tianshou.env.atari.atari_wrapper import make_atari_env
+from tianshou.env.atari import DQNet, make_atari_env
 from tianshou.highlevel.logger import LoggerFactoryDefault
 from tianshou.trainer import OfflineTrainerParams
 from tianshou.utils.net.discrete import DiscreteActor
