+++ examples/vizdoom/vizdoom_ppo.py
@@ -15,7 +15,7 @@ from tianshou.algorithm.modelbased.icm import ICMOnPolicyWrapper
 from tianshou.algorithm.modelfree.reinforce import ProbabilisticActorPolicy
 from tianshou.algorithm.optim import AdamOptimizerFactory, LRSchedulerFactoryLinear
 from tianshou.data import Collector, CollectStats, VectorReplayBuffer
-from tianshou.env.atari.atari_network import DQNet
+from tianshou.env.atari import DQNet
 from tianshou.highlevel.logger import LoggerFactoryDefault
 from tianshou.trainer import OnPolicyTrainerParams
 from tianshou.utils.net.discrete import (
